requirements.txt
telegram
python-telegram-bot==13.7
urllib3==1.26.18
requests
firebase-admin
bs4
pillow
datetime